+++ tests/snowpark/test_package.py
@@ -2,12 +2,13 @@ import logging
 import os
 from unittest import mock
 from unittest.mock import MagicMock, patch
-from zipfile import ZipFile
 
 import pytest
 from snowflake.cli.plugins.snowpark.models import (
     Requirement,
-    SplitRequirements,
+)
+from snowflake.cli.plugins.snowpark.package_utils import (
+    DownloadUnavailablePackagesResult,
 )
 
 from tests.test_data import test_data
@@ -31,56 +32,31 @@ class TestPackage:
         assert result.exit_code == 0
         assert result.output == snapshot
 
-    @patch("snowflake.cli.plugins.snowpark.package_utils.download_packages")
     @patch(
-        "snowflake.cli.plugins.snowpark.package.anaconda.AnacondaChannel.parse_anaconda_packages"
+        "snowflake.cli.plugins.snowpark.package.commands.download_unavailable_packages"
     )
-    def test_package_lookup_with_install_packages(
-        self, mock_package, mock_download, runner, capfd, snapshot
-    ) -> None:
-
-        mock_package.return_value = SplitRequirements(
-            [], [Requirement("some-other-package")]
-        )
-        mock_download.return_value = (
-            True,
-            SplitRequirements(
-                [Requirement("snowflake-snowpark-python")],
-                [Requirement("some-other-package")],
-            ),
-        )
-
-        result = runner.invoke(["snowpark", "package", "lookup", "some-other-package"])
-        assert result.exit_code == 0
-        assert result.output == snapshot
-
-    @patch("snowflake.cli.plugins.snowpark.package.commands.download_packages")
     @patch(
-        "snowflake.cli.plugins.snowpark.package.commands.AnacondaChannel.from_snowflake"
+        "snowflake.cli.plugins.snowpark.package_utils.Venv.pip_wheel",
     )
     @pytest.mark.parametrize(
         "extra_flags", [[], ["--skip-version-check"], ["--ignore-anaconda"]]
     )
     def test_package_create(
         self,
-        mock_anaconda_from_snowflake,
+        mock_pip_wheel,
         mock_download,
         caplog,
         temp_dir,
-        dot_packages_directory,
         runner,
         extra_flags,
     ) -> None:
+        mock_pip_wheel.return_value = 9
 
-        mock_anaconda = MagicMock(name="anaconda")
-        mock_anaconda.is_package_available.return_value = False
-        mock_anaconda_from_snowflake.return_value = mock_anaconda
-        mock_download.return_value = (
-            True,
-            SplitRequirements(
-                [Requirement.parse("in-anaconda-package>=2")],
-                [Requirement.parse("some-other-package")],
-            ),
+        mock_download.return_value = DownloadUnavailablePackagesResult(
+            succeeded=True,
+            packages_available_in_anaconda=[
+                Requirement.parse("in-anaconda-package>=2")
+            ],
         )
 
         with caplog.at_level(
@@ -93,29 +69,7 @@ class TestPackage:
 
         assert result.exit_code == 0, result.output
         assert "in-anaconda-package>=2" in result.output
-        assert "some-other-package" not in result.output
         assert os.path.isfile("totally-awesome-package.zip"), result.output
-
-        zip_file = ZipFile("totally-awesome-package.zip", "r")
-        assert (
-            "totally-awesome-package/totally-awesome-module.py" in zip_file.namelist()
-        )
-        os.remove("totally-awesome-package.zip")
-
-        if "--ignore-anaconda" in extra_flags:
-            expected_anaconda = expected_anaconda_from_snowfake = []
-        else:
-            expected_anaconda_from_snowfake = [mock.call()]
-            expected_anaconda = [
-                mock.call.is_package_available(
-                    Requirement.parse("totally-awesome-package"),
-                    skip_version_check="--skip-version-check" in extra_flags,
-                )
-            ]
-        assert mock_anaconda.mock_calls == expected_anaconda
-        assert (
-            mock_anaconda_from_snowflake.mock_calls == expected_anaconda_from_snowfake
-        )
 
     @mock.patch("snowflake.cli.plugins.snowpark.package.manager.StageManager")
     @mock.patch("snowflake.connector.connect")
@@ -181,7 +135,7 @@ class TestPackage:
             ["--pypi-download", "-y"],
         ],
     )
-    @mock.patch(
+    @patch(
         "snowflake.cli.plugins.snowpark.package.commands.AnacondaChannel.from_snowflake"
     )
     def test_lookup_install_flag_are_deprecated(self, _, flags, runner):
@@ -191,10 +145,10 @@ class TestPackage:
             in result.output
         )
 
-    @mock.patch(
+    @patch(
         "snowflake.cli.plugins.snowpark.package.commands.AnacondaChannel.from_snowflake"
     )
-    def test_lookup_install_with_out_flags_does_not_warn(self, _, runner):
+    def test_lookup_install_without_flags_does_not_warn(self, _, runner):
         result = runner.invoke(["snowpark", "package", "lookup", "foo"])
         assert (
             "is deprecated. Lookup command no longer checks for package in PyPi"
@@ -211,9 +165,15 @@ class TestPackage:
         ],
     )
     @mock.patch(
-        "snowflake.cli.plugins.snowpark.package.commands.AnacondaChannel.from_snowflake"
+        "snowflake.cli.plugins.snowpark.package.commands.download_unavailable_packages"
     )
-    def test_create_install_flag_are_deprecated(self, _, flags, runner):
+    @mock.patch("snowflake.cli.plugins.snowpark.package.commands.zip_dir")
+    @mock.patch(
+        "snowflake.cli.plugins.snowpark.package.commands.get_package_name_from_pip_wheel"
+    )
+    def test_create_install_flag_are_deprecated(
+        self, _mock_pip_wheel, _mock_zip, _mock_download, flags, runner
+    ):
         result = runner.invoke(["snowpark", "package", "create", "foo", *flags])
         assert (
             "is deprecated. Create command always checks for package in PyPi."
@@ -229,16 +189,28 @@ class TestPackage:
         ],
     )
     @mock.patch(
-        "snowflake.cli.plugins.snowpark.package.commands.AnacondaChannel.from_snowflake"
+        "snowflake.cli.plugins.snowpark.package.commands.download_unavailable_packages"
     )
-    def test_create_deprecated_flags_throw_warning(self, _, flags, runner):
+    @mock.patch("snowflake.cli.plugins.snowpark.package.commands.zip_dir")
+    @mock.patch(
+        "snowflake.cli.plugins.snowpark.package.commands.get_package_name_from_pip_wheel"
+    )
+    def test_create_deprecated_flags_throw_warning(
+        self, _mock_pip_wheel, _mock_zip, _mock_download, flags, runner
+    ):
         result = runner.invoke(["snowpark", "package", "create", "foo", *flags])
         assert "is deprecated." in result.output
 
     @mock.patch(
-        "snowflake.cli.plugins.snowpark.package.commands.AnacondaChannel.from_snowflake"
+        "snowflake.cli.plugins.snowpark.package.commands.download_unavailable_packages"
     )
-    def test_create_with_out_flags_does_not_warn(self, _, runner):
+    @mock.patch("snowflake.cli.plugins.snowpark.package.commands.zip_dir")
+    @mock.patch(
+        "snowflake.cli.plugins.snowpark.package.commands.get_package_name_from_pip_wheel"
+    )
+    def test_create_without_flags_does_not_warn(
+        self, _mock_pip_wheel, _mock_zip, _mock_download, runner
+    ):
         result = runner.invoke(["snowpark", "package", "create", "foo"])
         assert "is deprecated" not in result.output